+++ app/src/main/java/org/ieee/ieeehub/provider/article/ArticleModel.java
@@ -38,7 +38,7 @@ public interface ArticleModel extends BaseModel {
      * Can be {@code null}.
      */
     @Nullable
-    Integer getCategoryId();
+    Long getCategoryId();
 
     /**
      * Get the {@code image} value.
